+++ lib/src/char_matcher/char_match.dart
@@ -4,13 +4,13 @@ import 'char_matcher.dart';
 
 @immutable
 class CharMatch implements Match {
-  const CharMatch(this.start, this.input, this.pattern);
+  const CharMatch(this.start, this.end, this.input, this.pattern);
 
   @override
   final int start;
 
   @override
-  int get end => start + 1;
+  final int end;
 
   @override
   final String input;
@@ -19,10 +19,10 @@ class CharMatch implements Match {
   final CharMatcher pattern;
 
   @override
-  String? group(int group) => this[group];
+  String? group(int group) => group == 0 ? input.substring(start, end) : null;
 
   @override
-  String? operator [](int group) => group == 0 ? input[start] : null;
+  String? operator [](int group) => this.group(group);
 
   @override
   List<String?> groups(List<int> groupIndices) =>