+++ lib/src/char_matcher/char_set.dart
@@ -2,5 +2,5 @@ import 'char_matcher.dart';
 import 'optimize.dart';
 import 'range.dart';
 
-CharMatcher fromCharSet(String chars) => optimize(
-    chars.codeUnits.map((codeUnit) => RangeCharMatcher(codeUnit, codeUnit)));
+CharMatcher fromCharSet(String chars) =>
+    optimize(chars.runes.map((value) => RangeCharMatcher(value, value)));