+++ lib/src/char_matcher/letter.dart
@@ -1,9 +1,0 @@
-import 'char_matcher.dart';
-
-class LetterCharMatcher extends CharMatcher {
-  const LetterCharMatcher();
-
-  @override
-  bool match(int value) =>
-      (65 <= value && value <= 90) || (97 <= value && value <= 122);
-}