+++ lib/src/char_matcher/lower_case.dart
@@ -1,8 +1,0 @@
-import 'char_matcher.dart';
-
-class LowerCaseLetterCharMatcher extends CharMatcher {
-  const LowerCaseLetterCharMatcher();
-
-  @override
-  bool match(int value) => 97 <= value && value <= 122;
-}