+++ lib/src/char_matcher/single.dart
@@ -2,14 +2,14 @@ import '../../printer.dart';
 import 'char_matcher.dart';
 
 class SingleCharMatcher extends CharMatcher {
-  const SingleCharMatcher(this.charValue);
+  const SingleCharMatcher(this.codePoint);
 
-  final int charValue;
+  final int codePoint;
 
   @override
-  bool match(int value) => identical(charValue, value);
+  bool match(int value) => identical(codePoint, value);
 
   @override
   ObjectPrinter get toStringPrinter =>
-      super.toStringPrinter..addValue(charValue, name: 'charValue');
+      super.toStringPrinter..addValue(codePoint, name: 'codePoint');
 }