+++ test/collection_test.dart
@@ -1,4 +1,4 @@
-// ignore_for_file: deprecated_member_use_from_same_package, unnecessary_lambdas
+// ignore_for_file: deprecated_member_use_from_same_package, unnecessary_lambdas, collection_methods_unrelated_type
 
 import 'dart:collection';
 import 'dart:math';